Move general federation servlets to a separate file.
